refactored anonSignin, linkUsers, remove netMgr
