+++ Potluck/Potluck/Networking/MHPRequestHandler.swift
@@ -30,6 +30,28 @@ struct MHPRequestHandler {
         }
     }
     
+    func signInAnon(completion: @escaping (Result<MHPUser, Error> ) -> ()) {
+        service.router.signInAnon { (result) in
+            switch result {
+            case .success (let mhpUser):
+                completion(.success(mhpUser))
+            case .failure(let error):
+                completion(.failure(error))
+            }
+        }
+    }
+    
+    func linkUsers(email: String, password: String, mhpUser: MHPUser, completion: @escaping (Result<MHPUser, Error> ) -> ()) {
+        service.router.linkUsers(email: email, password: password, mhpUser: mhpUser) { (result) in
+            switch result {
+            case .success (let mhpUser):
+                completion(.success(mhpUser))
+            case .failure(let error):
+                completion(.failure(error))
+            }
+        }
+    }
+    
     func retrieveUser(completion: @escaping (Result<MHPUser, Error> ) -> ()) {
         service.router.retrieveUser { (result) in
             switch result {
@@ -79,6 +101,8 @@ struct MHPRequestHandler {
 /// Sends a request from the RequestHandler to the selected service
 class MHPServiceRouter: Routeable {
     func loginUser(email: String, password: String, completion: @escaping (Result<MHPUser, Error> ) -> ()) {}
+    func signInAnon(completion: @escaping (Result<MHPUser, Error> ) -> ()) {}
+    func linkUsers(email: String, password: String, mhpUser: MHPUser, completion: @escaping (Result<MHPUser, Error> ) -> ()) {}
     func retrieveUser(completion: @escaping (Result<MHPUser, Error> ) -> ()) {}
     func updateUserState(mhpUser: MHPUser, state: UserAuthorizationState, completion: @escaping (Result<Bool, Error> ) -> ()) {}
     func sendVerificationEmail(completion: @escaping (Result<Bool, Error> ) -> ()) {}
@@ -92,11 +116,15 @@ protocol Routeable {
 class MHPFirebaseFirestoreServiceRouter: MHPServiceRouter {
     let db = Firestore.firestore()
     let dataManager = MHPDataManager()
+    let actionCodeSettings = ActionCodeSettings.init()
     
     override init() {
         let settings = db.settings
         settings.areTimestampsInSnapshotsEnabled = true
         db.settings = settings
+        
+        actionCodeSettings.handleCodeInApp = true
+        actionCodeSettings.setIOSBundleID(Bundle.main.bundleIdentifier!)
     }
     
     override func loginUser(email: String, password: String, completion: @escaping (Result<MHPUser, Error> ) -> ()) {
@@ -113,6 +141,75 @@ class MHPFirebaseFirestoreServiceRouter: MHPServiceRouter {
             } else {
                 completion(.failure(error!))
             }
+        }
+    }
+    
+    override func signInAnon(completion: @escaping (Result<MHPUser, Error> ) -> ()) {
+        Auth.auth().signInAnonymously() { (user, error) in
+            if error == nil {
+                if let returnedUser = user {
+                    let ref: DocumentReference = self.db.collection("users").document(returnedUser.uid)
+                    let dataSet = self.dataManager.buildDataSet(firUserEmail: returnedUser.email, mhpUser: nil, firstName: nil, lastName: nil, state: .anonymous)
+                    ref.setData(dataSet, options: SetOptions.merge()) { (error) in
+                        if let error = error {
+                            print("Error adding document: \(error)")
+                            completion(.failure(error))
+                        } else {
+                            print("Anon user added with ID: \(ref.documentID)")
+                            // retrieve mhpUser from db
+                            self.retrieveUser(completion: { (result) in
+                                switch result {
+                                case .success(let mhpUser):
+                                    completion(.success(mhpUser))
+                                default:
+                                    completion(.failure(error!))
+                                }
+                            })
+                        }
+                        
+                    }
+                }
+            } else {
+                completion(.failure(error!))
+            }
+        }
+    }
+    
+    override func linkUsers(email: String, password: String, mhpUser: MHPUser, completion: @escaping (Result<MHPUser, Error> ) -> ()) {
+        // link newly created user to anon user in Firestore
+        let credential = EmailAuthProvider.credential(withEmail: email, password: password)
+        if let currentUser = Auth.auth().currentUser {
+            currentUser.link(with: credential, completion: { (user, error) in
+                if error == nil {
+                    self.sendVerificationEmail { (result) in
+                        switch result {
+                        case .success:
+                            return
+                        case .failure:
+                            completion(.failure(error!))
+                        }
+                    }
+                    // save updated mhpUser info to DB
+                    self.updateUserState(mhpUser: mhpUser, state: .unverified, completion: { (result) in
+                        switch result {
+                        case .success(_):
+                            // retrieve mhpUser from db
+                            self.retrieveUser(completion: { (result) in
+                                switch result {
+                                case .success(let mhpUser):
+                                    completion(.success(mhpUser))
+                                default:
+                                    completion(.failure(error!))
+                                }
+                            })
+                        default:
+                            completion(.failure(error!))
+                        }
+                    })
+                } else {
+                    completion(.failure(error!))
+                }
+            })
         }
     }
     
@@ -169,11 +266,8 @@ class MHPFirebaseFirestoreServiceRouter: MHPServiceRouter {
     }
     
     override func sendVerificationEmail(completion: @escaping (Result<Bool, Error> ) -> ()) {
-        let actionCodeSettings =  ActionCodeSettings.init()
-        actionCodeSettings.handleCodeInApp = true
         if let user = Auth.auth().currentUser, let email = user.email {
             actionCodeSettings.url = URL(string: "https://tza3e.app.goo.gl/emailVerification/?email=\(email)")
-            actionCodeSettings.setIOSBundleID(Bundle.main.bundleIdentifier!)
             user.sendEmailVerification(completion: { (error) in
                 if error == nil {
                     print("verification email sent")
@@ -188,10 +282,7 @@ class MHPFirebaseFirestoreServiceRouter: MHPServiceRouter {
     }
     
     override func sendResetPasswordEmail(forEmail email: String, completion: @escaping (Result<Bool, Error> ) -> ()) {
-        let actionCodeSettings =  ActionCodeSettings.init()
-        actionCodeSettings.handleCodeInApp = true
         actionCodeSettings.url = URL(string: "https://tza3e.app.goo.gl/resetPassword/?email=\(email)")
-        actionCodeSettings.setIOSBundleID(Bundle.main.bundleIdentifier!)
         Auth.auth().sendPasswordReset(withEmail: email) { (error) in
             if error == nil {
                 completion(.success(true))
